Getting the folder sharing is a hassle for testing
Moved the folder sharing to the end of the process, as the /etc/hosts
fix needs to have run before the folder sharing is activated. Activating
the folder sharing by default requires a more complex setup, and makes
testing a problem, so I have deactivated the folder sharing.

Added the option to specify the memory size and number of CPUs for the
VirtualBox. Newer versions of Ubuntu work better with more RAM taking
into account the software installed in the VirtualBox.