Added a vbox name and linked_clone
If more than one vagrant is started, it makes sense to link the clone so
that there is no need to always download and install the base image.
If more than one vagrant is installed , then the name ought to be
different. If a name is not specified, then the name is that of the
image, so if one has more than one vagrant setup that uses the same base
image, without a name only one image could be installed.